Avatar refactoring (default + config)
